"Add member group creation UI and layout updates
- Extract create-group logic into reusable GroupCreationForm component
- Wire Groups route under member dashboard
- Clean up selected-members display and button enabling logic"
